[#6763] Minimize all open windows during placement
Finds all applications that are not minimized, have a frame, and
are not detached and minimizes them until placement is completed.
This behavior can be disabled using the `minimizeWindows` option
in the placement configuration.

Also adds a `restoreLayer` option to the configuration to control
whether the original canvas layer is made active after placement.
Both of these options are enabled by default.

Closes #6763
