+++ module/canvas/_module.mjs
@@ -1,3 +1,4 @@
+export {default as BasePlacement} from "./api/base-placement.mjs";
 export * as detectionModes from "./detection-modes/_module.mjs";
 export * as layers from "./layers/_module.mjs";
 export {default as MapLocationControlIcon} from "./map-location-control-icon.mjs";
