+++ module/data/actor/templates/group.mjs
@@ -67,8 +67,6 @@ export default class GroupTemplate extends ActorDataModel.mixin(CurrencyTemplate
     if ( !game.user.isGM || !canvas.scene ) return [];
     const members = await this.getPlaceableMembers();
     if ( !members.length ) return [];
-    const minimized = !this.parent.sheet.minimized;
-    await this.parent.sheet.minimize();
     const tokensData = [];
 
     try {
@@ -91,8 +89,6 @@ export default class GroupTemplate extends ActorDataModel.mixin(CurrencyTemplate
         log: "error",
         notify: "error"
       });
-    } finally {
-      if ( minimized ) this.parent.sheet.maximize();
     }
 
     return canvas.scene.createEmbeddedDocuments("Token", tokensData);
